Implement Perft debugging tool for counting nodes at certain depths and bug fixes.

+++ src/board/zobrist.h
@@ -52,7 +52,7 @@ public:
         }())
     {}
 
-    uint32_t getKey() const { return key; };
+    uint64_t getKey() const { return key; };
 };
 
 #endif //TEMPO_ZOBRIST_H
